requirements.txt
fastapi==0.110.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
uvicorn==0.27.1
